Prevent underflow foreal, thanks nishant!

+++ beacon-chain/db/kv/blocks.go
@@ -317,7 +317,7 @@ func (k *Store) HighestSlotBlocks(ctx context.Context) ([]*ethpb.SignedBeaconBlo
 			return err
 		}
 
-		blocks, err = k.blocksAtSlotBitfieldIndex(ctx, tx, uint64(highestIndex))
+		blocks, err = k.blocksAtSlotBitfieldIndex(ctx, tx, highestIndex)
 		if err != nil {
 			return err
 		}
@@ -340,7 +340,7 @@ func (k *Store) HighestSlotBlocksBelow(ctx context.Context, slot uint64) ([]*eth
 		if err != nil {
 			return err
 		}
-		blocks, err = k.blocksAtSlotBitfieldIndex(ctx, tx, uint64(highestIndex))
+		blocks, err = k.blocksAtSlotBitfieldIndex(ctx, tx, highestIndex)
 		if err != nil {
 			return err
 		}
@@ -352,13 +352,13 @@ func (k *Store) HighestSlotBlocksBelow(ctx context.Context, slot uint64) ([]*eth
 
 // blocksAtSlotBitfieldIndex retrieves the blocks in DB given the input index. The index represents
 // the position of the slot bitfield the saved block maps to.
-func (k *Store) blocksAtSlotBitfieldIndex(ctx context.Context, tx *bolt.Tx, index uint64) ([]*ethpb.SignedBeaconBlock, error) {
+func (k *Store) blocksAtSlotBitfieldIndex(ctx context.Context, tx *bolt.Tx, index int) ([]*ethpb.SignedBeaconBlock, error) {
 	ctx, span := trace.StartSpan(ctx, "BeaconDB.blocksAtSlotBitfieldIndex")
 	defer span.End()
 
 	highestSlot := index - 1
-	highestSlot = uint64(math.Max(0, float64(highestSlot)))
-	f := filters.NewFilter().SetStartSlot(highestSlot).SetEndSlot(highestSlot)
+	highestSlot = int(math.Max(0, float64(highestSlot)))
+	f := filters.NewFilter().SetStartSlot(uint64(highestSlot)).SetEndSlot(uint64(highestSlot))
 
 	keys, err := getBlockRootsByFilter(ctx, tx, f)
 	if err != nil {
